+++ packages/frameworks/solid/src/date-picker/use-date-picker.tsx
@@ -4,18 +4,27 @@ import { createMemo, createUniqueId, splitProps, type Accessor } from 'solid-js'
 import { useEnvironmentContext } from '../environment'
 import { type Optional } from '../types'
 
-export interface UseDatePickerProps extends Optional<Omit<datePicker.Context, 'value'>, 'id'> {
+export interface UseDatePickerProps
+  extends Optional<Omit<datePicker.Context, 'value' | 'focusedValue'>, 'id'> {
+  /**
+   * The focused date.
+   */
+  focusedValue?: string
+  /**
+   * The value of the date picker
+   */
   value?: string[]
 }
 export interface UseDatePickerReturn extends Accessor<datePicker.Api<PropTypes>> {}
 
 export const useDatePicker = (props: UseDatePickerProps): UseDatePickerReturn => {
-  const [local, rest] = splitProps(props, ['value'])
+  const [local, rest] = splitProps(props, ['value', 'focusedValue'])
   const getRootNode = useEnvironmentContext()
   const context = mergeProps(
     () => ({
       id: createUniqueId(),
       getRootNode,
+      focusedValue: local.focusedValue ? datePicker.parse(local.focusedValue) : undefined,
       value: local.value ? datePicker.parse(local.value) : undefined,
     }),
     rest,